+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/control/DefaultControlKey.java
@@ -33,9 +33,7 @@ final class DefaultControlKey<T extends Control> implements ControlKey<T> {
 	DefaultControlKey(String name, Class<T> controlClass, KeyStroke defaultKeyStroke) {
 		this.name = requireNonNull(name);
 		this.controlClass = requireNonNull(controlClass);
-		this.defaultKeyStroke = Value.builder()
-						.nullable(defaultKeyStroke)
-						.build();
+		this.defaultKeyStroke = Value.value(defaultKeyStroke);
 	}
 
 	@Override