+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/control/DefaultControlsKey.java
@@ -35,9 +35,7 @@ final class DefaultControlsKey implements ControlsKey {
 
 	DefaultControlsKey(String name, KeyStroke defaultKeyStroke, Layout defaultLayout) {
 		this.name = requireNonNull(name);
-		this.defaultKeyStroke = Value.builder()
-						.nullable(defaultKeyStroke)
-						.build();
+		this.defaultKeyStroke = Value.value(defaultKeyStroke);
 		this.defaultLayout = defaultLayout;
 	}