+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/AbstractDialogBuilder.java
@@ -78,9 +78,7 @@ public class AbstractDialogBuilder<B extends DialogBuilder<B>> implements Dialog
 
 	@Override
 	public final B title(String title) {
-		return title(Value.builder()
-						.nullable(title)
-						.build());
+		return title(Value.value(title));
 	}
 
 	@Override