+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/ComponentsTest.java
@@ -113,9 +113,7 @@ public final class ComponentsTest {
 	void clear() {
 		Font defaultFont = new JTextField().getFont();
 
-		Value<Integer> value = Value.builder()
-						.nullable(42)
-						.build();
+		Value<Integer> value = Value.value(42);
 
 		NumberField.Builder<Integer> builder = Components.integerField()
 						.valueRange(0, 100)
@@ -162,7 +160,7 @@ public final class ComponentsTest {
 
 	@Test
 	void shortField() {
-		Value<Short> value = Value.builder().nullable((short) 42).build();
+		Value<Short> value = Value.value((short) 42);
 		ComponentValue<Short, NumberField<Short>> componentValue = Components.shortField()
 						.valueRange(0, 100)
 						.font(Font.getFont("arial"))
@@ -177,7 +175,7 @@ public final class ComponentsTest {
 
 	@Test
 	void integerField() {
-		Value<Integer> value = Value.builder().nullable(42).build();
+		Value<Integer> value = Value.value(42);
 		ComponentValue<Integer, NumberField<Integer>> componentValue = Components.integerField()
 						.valueRange(0, 100)
 						.font(Font.getFont("arial"))
@@ -192,7 +190,7 @@ public final class ComponentsTest {
 
 	@Test
 	void longField() {
-		Value<Long> value = Value.builder().nullable(42L).build();
+		Value<Long> value = Value.value(42L);
 		ComponentValue<Long, NumberField<Long>> componentValue = Components.longField()
 						.valueRange(0, 100)
 						.groupingSeparator('.')
@@ -205,7 +203,7 @@ public final class ComponentsTest {
 
 	@Test
 	void doubleField() {
-		Value<Double> value = Value.builder().nullable(42.2).build();
+		Value<Double> value = Value.value(42.2);
 		ComponentValue<Double, NumberField<Double>> componentValue = Components.doubleField()
 						.valueRange(0, 100)
 						.maximumFractionDigits(2)
@@ -220,7 +218,7 @@ public final class ComponentsTest {
 
 	@Test
 	void bigDecimalField() {
-		Value<BigDecimal> value = Value.builder().nullable(BigDecimal.valueOf(42.2)).build();
+		Value<BigDecimal> value = Value.value(BigDecimal.valueOf(42.2));
 		ComponentValue<BigDecimal, NumberField<BigDecimal>> componentValue = Components.bigDecimalField()
 						.maximumFractionDigits(2)
 						.groupingSeparator('.')
@@ -233,7 +231,7 @@ public final class ComponentsTest {
 
 	@Test
 	void localTimeField() {
-		Value<LocalTime> value = Value.builder().nullable(LocalTime.now()).build();
+		Value<LocalTime> value = Value.value(LocalTime.now());
 		ComponentValue<LocalTime, TemporalField<LocalTime>> componentValue =
 						Components.localTimeField("HH:mm")
 										.focusLostBehaviour(JFormattedTextField.COMMIT)
@@ -244,7 +242,7 @@ public final class ComponentsTest {
 
 	@Test
 	void localDateField() {
-		Value<LocalDate> value = Value.builder().nullable(LocalDate.now()).build();
+		Value<LocalDate> value = Value.value(LocalDate.now());
 		ComponentValue<LocalDate, TemporalField<LocalDate>> componentValue =
 						Components.localDateField("dd-MM-yyyy")
 										.focusLostBehaviour(JFormattedTextField.COMMIT)
@@ -255,7 +253,7 @@ public final class ComponentsTest {
 
 	@Test
 	void localDateTimeField() {
-		Value<LocalDateTime> value = Value.builder().nullable(LocalDateTime.now()).build();
+		Value<LocalDateTime> value = Value.value(LocalDateTime.now());
 		ComponentValue<LocalDateTime, TemporalField<LocalDateTime>> componentValue =
 						Components.localDateTimeField("dd-MM-yyyy HH:mm")
 										.focusLostBehaviour(JFormattedTextField.COMMIT)
@@ -266,7 +264,7 @@ public final class ComponentsTest {
 
 	@Test
 	void offsetDateTimeField() {
-		Value<OffsetDateTime> value = Value.builder().nullable(OffsetDateTime.now()).build();
+		Value<OffsetDateTime> value = Value.value(OffsetDateTime.now());
 		ComponentValue<OffsetDateTime, TemporalField<OffsetDateTime>> componentValue =
 						Components.offsetDateTimeField("yyyy-MM-dd'T'HH:mm:ss.SSSZ")
 										.focusLostBehaviour(JFormattedTextField.COMMIT)
@@ -514,7 +512,7 @@ public final class ComponentsTest {
 
 	@Test
 	void nullableCheckBox() {
-		Value<Boolean> value = Value.builder().nullable(true).build();
+		Value<Boolean> value = Value.value(true);
 		ComponentValue<Boolean, JCheckBox> componentValue = Components.checkBox(value)
 						.transferFocusOnEnter(true)
 						.nullable(true)
@@ -534,7 +532,7 @@ public final class ComponentsTest {
 
 	@Test
 	void booleanComboBox() {
-		Value<Boolean> value = Value.builder().nullable(true).build();
+		Value<Boolean> value = Value.value(true);
 		ComponentValue<Boolean, JComboBox<Item<Boolean>>> componentValue =
 						Components.booleanComboBox(booleanItemComboBoxModel())
 										.maximumRowCount(5)
@@ -652,7 +650,7 @@ public final class ComponentsTest {
 		textArea = builder.build();
 		assertEquals(textArea.getDocument().getLength(), textArea.getCaretPosition());
 
-		Value<String> stringValue = Value.builder().nullable("hello there").build();
+		Value<String> stringValue = Value.value("hello there");
 		builder = Components.textArea(stringValue)
 						.initialCaretPosition(InitialCaretPosition.START);
 		textArea = builder.build();
@@ -723,7 +721,7 @@ public final class ComponentsTest {
 
 	@Test
 	void integerSpinner() {
-		Value<Integer> value = Value.builder().nullable(10).build();
+		Value<Integer> value = Value.value(10);
 		ComponentValue<Integer, JSpinner> componentValue = Components.integerSpinner()
 						.minimum(0)
 						.maximum(100)
@@ -739,7 +737,7 @@ public final class ComponentsTest {
 
 	@Test
 	void doubleSpinner() {
-		Value<Double> value = Value.builder().nullable(10d).build();
+		Value<Double> value = Value.value(10d);
 		ComponentValue<Double, JSpinner> componentValue = Components.doubleSpinner()
 						.minimum(0d)
 						.maximum(100d)
@@ -788,7 +786,7 @@ public final class ComponentsTest {
 
 	@Test
 	void slider() {
-		Value<Integer> value = Value.builder().nullable(10).build();
+		Value<Integer> value = Value.value(10);
 		ComponentValue<Integer, JSlider> componentValue = Components.slider(new DefaultBoundedRangeModel(0, 0, 0, 100))
 						.snapToTicks(true)
 						.paintTrack(true)
@@ -808,7 +806,7 @@ public final class ComponentsTest {
 
 	@Test
 	void label() {
-		Value<String> textValue = Value.builder().nullable("label").build();
+		Value<String> textValue = Value.value("label");
 		ComponentValue<String, JLabel> componentValue = Components.label(textValue)
 						.icon(Logos.logoTransparent())
 						.iconTextGap(5)
@@ -853,7 +851,7 @@ public final class ComponentsTest {
 		listModel.addElement("two");
 		listModel.addElement("three");
 
-		Value<String> textValue = Value.builder().nullable("two").build();
+		Value<String> textValue = Value.value("two");
 		ListBuilder.SelectedItem<String> listBuilder = Components.list(listModel)
 						.selectedItem()
 						.visibleRowCount(4)
@@ -934,7 +932,7 @@ public final class ComponentsTest {
 						.validator(validator)
 						.build());
 
-		Value<String> stringValue = Value.builder().nullable("test").build();
+		Value<String> stringValue = Value.value("test");
 		assertThrows(IllegalArgumentException.class, () -> Components.textField(String.class, stringValue)
 						.validator(validator)
 						.build());